refactor(ui): split mixed files and fix a locale-dependent HAL check
