+++ app/src/main/java/com/aoscoremonitor/ui/components/EmptyState.kt
@@ -1,9 +1,6 @@
 package com.aoscoremonitor.ui.components
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -18,7 +15,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.aoscoremonitor.ui.theme.Spacing
 
@@ -28,6 +24,8 @@ import com.aoscoremonitor.ui.theme.Spacing
  * Sits inline where the cards would have been, so the surrounding headings and tabs stay put.
  * Replaces the per-screen `EmptyStateMessage` and `UnavailableNotice`, which were the same
  * component under two names.
+ *
+ * Use [FullScreenMessage] when the whole screen is empty.
  */
 @Composable
 fun EmptyState(message: String, modifier: Modifier = Modifier, icon: ImageVector = Icons.Default.Info) {
@@ -52,45 +50,6 @@ fun EmptyState(message: String, modifier: Modifier = Modifier, icon: ImageVector
                 style = MaterialTheme.typography.bodyMedium,
                 color = MaterialTheme.colorScheme.onSurfaceVariant,
                 modifier = Modifier.padding(start = Spacing.Medium)
-            )
-        }
-    }
-}
-
-/**
- * The same explanation when it is the only thing on the screen.
- *
- * Centred rather than pinned to the top, because a lone card floating under the app bar reads as
- * a loading glitch.
- */
-@Composable
-fun FullScreenMessage(message: String, icon: ImageVector, modifier: Modifier = Modifier, supportingText: String? = null) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(Spacing.ExtraLarge),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.size(48.dp)
-        )
-        Text(
-            text = message,
-            style = MaterialTheme.typography.titleMedium,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(top = Spacing.Large)
-        )
-        if (supportingText != null) {
-            Text(
-                text = supportingText,
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                textAlign = TextAlign.Center,
-                modifier = Modifier.padding(top = Spacing.Small)
             )
         }
     }
