+++ app/src/main/java/com/aoscoremonitor/ui/components/FullScreenMessage.kt
@@ -1,0 +1,59 @@
+package com.aoscoremonitor.ui.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.aoscoremonitor.ui.theme.Spacing
+
+/**
+ * The only thing on a screen, when there is nothing else to show.
+ *
+ * Centred rather than pinned to the top, because a lone card floating under the app bar reads as
+ * a loading glitch. Named for the layout rather than for "empty" because the network and TCP
+ * screens also use it to say they are still reading.
+ *
+ * Use [EmptyState] when only one section of a screen is empty.
+ */
+@Composable
+fun FullScreenMessage(message: String, icon: ImageVector, modifier: Modifier = Modifier, supportingText: String? = null) {
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(Spacing.ExtraLarge),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Icon(
+            imageVector = icon,
+            contentDescription = null,
+            tint = MaterialTheme.colorScheme.onSurfaceVariant,
+            modifier = Modifier.size(48.dp)
+        )
+        Text(
+            text = message,
+            style = MaterialTheme.typography.titleMedium,
+            textAlign = TextAlign.Center,
+            modifier = Modifier.padding(top = Spacing.Large)
+        )
+        if (supportingText != null) {
+            Text(
+                text = supportingText,
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                textAlign = TextAlign.Center,
+                modifier = Modifier.padding(top = Spacing.Small)
+            )
+        }
+    }
+}
