+++ app/src/main/java/com/aoscoremonitor/ui/components/MonitorCard.kt
@@ -89,47 +89,3 @@ fun InfoCard(
         }
     }
 }
-
-/**
- * A label and its value stacked as one unit.
- *
- * Screens were writing `"Version: ${x}"` into a single [Text], which glues the label to the value
- * for a screen reader and leaves no way to style them apart. Keeping them separate fixes both.
- */
-@Composable
-fun LabeledValue(label: String, value: String, modifier: Modifier = Modifier, valueStyle: TextStyle = MaterialTheme.typography.bodyMedium) {
-    Column(modifier = modifier) {
-        Text(
-            text = label,
-            style = MaterialTheme.typography.labelMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Text(text = value, style = valueStyle)
-    }
-}
-
-/**
- * A row stating one thing and whether it is in good shape.
- *
- * The status icon carries a spoken description because the color on its own conveys nothing to
- * TalkBack or to a colorblind user.
- */
-@Composable
-fun StatusRow(label: String, status: ReadingStatus, statusDescription: String, modifier: Modifier = Modifier) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            imageVector = status.icon,
-            contentDescription = statusDescription,
-            tint = status.color,
-            modifier = Modifier.size(20.dp)
-        )
-        Text(
-            text = label,
-            style = MaterialTheme.typography.bodyMedium,
-            modifier = Modifier.padding(start = Spacing.Small)
-        )
-    }
-}
